travel-rag-bot/requirements.txt
flask
langchain
langchain-community
langchain-huggingface
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
sentence-transformers